asset_daemon fixes for both native and docker

+++ native/database/nosql/mongo/mongo.cpp
@@ -36,8 +36,8 @@ std::optional<uint8_p> Mongo::fileDownloadWithId(MongoBucket &bucket, const std:
 
     auto downloadStream = bucket().open_download_stream(MongoObjectId{strview{_id}}());
     auto buffer = make_uint8_p(downloadStream.file_length());
-    auto ret = static_cast<size_t >(downloadStream.read(buffer.first.get(), downloadStream.file_length()));
-    if (ret == downloadStream.file_length()) {
+    auto ret = downloadStream.read(buffer.first.get(), downloadStream.file_length());
+    if (ret == static_cast<size_t >(downloadStream.file_length())) {
         return buffer;
     }
     return std::nullopt;
@@ -56,7 +56,7 @@ std::string Mongo::fileDownload(MongoBucket &bucket, const MongoObjectId &_id, c
     auto downloadStream = bucket().open_download_stream(_id());
     auto buffer = make_uint8_p(downloadStream.file_length());
     auto ret = static_cast<size_t >(downloadStream.read(buffer.first.get(), downloadStream.file_length()));
-    if (ret == downloadStream.file_length()) {
+    if (ret == static_cast<size_t >(downloadStream.file_length())) {
         FM::writeLocalFile(filename, reinterpret_cast<const char *>(buffer.first.get()), buffer.second, true);
         return filename;
     }
